mordecai stuff

+++ src/stage.h
@@ -251,7 +251,7 @@ typedef struct
 	s16 noteshakex;
 	s16 noteshakey;
 
-	s16 fade;
+	int fade;
 	s16 song_step;
 
 	s16 botplay;
